+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/service/android/tile/VpnQuickTile.kt
@@ -19,7 +19,6 @@ import net.nymtech.vpn.backend.Tunnel
 import nym_vpn_lib_types.EntryPoint
 import nym_vpn_lib_types.ExitPoint
 import timber.log.Timber
-import java.util.*
 import javax.inject.Inject
 
 @AndroidEntryPoint