+++ nym-vpn-core/crates/nym-vpnd/src/logging.rs
@@ -21,7 +21,7 @@ use nym_vpn_lib_types::LogPath;
 
 use crate::service;
 
-static INFO_CRATES: &[&str; 13] = &[
+static INFO_CRATES: &[&str; 14] = &[
     "hyper",
     "netlink_proto",
     "hickory_proto",
@@ -29,6 +29,7 @@ static INFO_CRATES: &[&str; 13] = &[
     "hyper_util",
     "h2",
     "rustls",
+    "surge_ping::client",
     "nym_statistics_common",
     "nym_sphinx_chunking",
     "nym_sphinx::preparer",