+++ nym-vpn-core/crates/nym-vpnd/src/service/vpn_service.rs
@@ -4,8 +4,8 @@
 use std::{path::PathBuf, pin::Pin};
 
 use bip39::Mnemonic;
-use futures::{FutureExt, StreamExt, future::Fuse, pin_mut};
-use time::{OffsetDateTime, format_description::well_known::Rfc3339};
+use futures::{future::Fuse, pin_mut, FutureExt, StreamExt};
+use time::{format_description::well_known::Rfc3339, OffsetDateTime};
 use tokio::{
     sync::{broadcast, mpsc, oneshot},
     task::JoinHandle,
@@ -16,18 +16,18 @@ use tokio_util::sync::CancellationToken;
 
 use nym_common::trace_err_chain;
 use nym_statistics::{
-    StatisticsController, StatisticsControllerConfig,
-    events::{StatisticsEvent, StatisticsSender},
+    events::{StatisticsEvent, StatisticsSender}, StatisticsController,
+    StatisticsControllerConfig,
 };
 use nym_vpn_account_controller::{
     AccountCommandSender, AccountController, AccountControllerConfig, AccountStateReceiver,
     AvailableTicketbooks, NyxdClient,
 };
-use nym_vpn_api_client::{ResolverOverrides, api_urls_to_urls, fronted_http_client};
+use nym_vpn_api_client::{api_urls_to_urls, fronted_http_client};
 use nym_vpn_lib::{
-    UserAgent, VpnTopologyProvider,
-    gateway_directory::{self, GatewayCache, GatewayCacheHandle, GatewayClient},
-    tunnel_state_machine::{NymConfig, TunnelCommand, TunnelConstants, TunnelStateMachine},
+    gateway_directory::{self, GatewayCache, GatewayCacheHandle, GatewayClient}, tunnel_state_machine::{NymConfig, TunnelCommand, TunnelConstants, TunnelStateMachine},
+    UserAgent,
+    VpnTopologyProvider,
 };
 use nym_vpn_lib_types::{
     AccountBalanceResponse, AccountCommandError, AccountControllerState, ConnectArgs,
@@ -36,7 +36,7 @@ use nym_vpn_lib_types::{
     NymVpnDevice, NymVpnNetwork, NymVpnUsage, ParsedAccountLinks, StoreAccountRequest,
     SystemMessage, TargetState, TunnelEvent, TunnelState, VpnServiceConfig, VpnServiceInfo,
 };
-use nym_vpn_network_config::Network;
+use nym_vpn_network_config::{start_discovery_refresher, Network};
 use nym_vpn_store::types::{StorableAccount, StoredAccountMode};
 
 use super::{
@@ -206,6 +206,9 @@ pub struct NymVpnService {
 
     // Gateway cache handle
     gateway_cache_handle: GatewayCacheHandle,
+
+    // Discovery Refresher join handle
+    discovery_refresher_join_handle: JoinHandle<()>,
 
     // VPN service shutdown token.
     shutdown_token: CancellationToken,
@@ -407,7 +410,7 @@ impl NymVpnService {
                 })?;
 
         let nym_config = NymConfig {
-            config_path: Some(config_dir),
+            config_path: Some(config_dir.clone()),
             data_path: Some(network_data_dir.clone()),
             gateway_config: gateway_config.clone(),
             network_env: *parameters.network_env.clone(),
@@ -428,40 +431,54 @@ impl NymVpnService {
             services_shutdown_token.child_token(),
         );
 
-        let urls = api_urls_to_urls(&nym_api_urls).map_err(|e| {
-            AccountControllerError::Initialization {
-                reason: e.to_string(),
-            }
-        })?;
-
-        let resolver_overrides = ResolverOverrides::from_urls(&urls).await.map_err(|e| {
-            AccountControllerError::Initialization {
-                reason: e.to_string(),
-            }
-        })?;
-
-        let validator_client = fronted_http_client(
-            urls.clone(),
-            Some(parameters.user_agent.clone()),
-            None,
-            Some(&resolver_overrides),
-        )
-        .await
-        .map_err(|err| {
-            tracing::error!("Failed to create HTTP client: {err:?}");
+        let urls = api_urls_to_urls(&nym_api_urls).map_err(|err| {
+            tracing::error!("Failed to convert Nym API URLs: {err:?}");
             AccountControllerError::Initialization {
                 reason: err.to_string(),
             }
         })?;
 
+        let validator_client =
+            fronted_http_client(urls, Some(parameters.user_agent.clone()), None, None)
+                .await
+                .map_err(|err| {
+                    tracing::error!("Failed to create HTTP client: {err:?}");
+                    AccountControllerError::Initialization {
+                        reason: err.to_string(),
+                    }
+                })?;
+
+        let urls = parameters.network_env.nym_api_urls_as_urls().ok_or(
+            AccountControllerError::Initialization {
+                reason: "Nym API URLs are empty".to_string(),
+            },
+        )?;
+
         let topology_provider = VpnTopologyProvider::new(
             urls,
-            Some(parameters.user_agent.clone()),
+            None,
             validator_client,
             false,
             services_shutdown_token.child_token(),
         );
         topology_provider.fetch().await;
+
+        let (discovery_refresher_events_tx, discovery_refresher_events_rx) = mpsc::channel(1);
+        let (discovery_refresher_commands_tx, discovery_refresher_commands_rx) = mpsc::channel(1);
+        let discovery_refresher_join_handle = start_discovery_refresher(
+            config_dir.clone(),
+            *parameters.network_env.clone(),
+            discovery_refresher_commands_rx,
+            discovery_refresher_events_tx,
+            services_shutdown_token.child_token(),
+        )
+        .await
+        .map_err(|err| {
+            tracing::error!("Failed to start Discovery Refresher: {err:?}");
+            AccountControllerError::Initialization {
+                reason: err.to_string(),
+            }
+        })?;
 
         let state_machine_handle = TunnelStateMachine::spawn(
             command_receiver,
@@ -475,6 +492,8 @@ impl NymVpnService {
             gateway_cache_handle.clone(),
             topology_provider,
             connectivity_handle,
+            discovery_refresher_commands_tx,
+            discovery_refresher_events_rx,
             #[cfg(not(any(target_os = "android", target_os = "ios")))]
             route_handler,
             state_machine_shutdown_token.child_token(),
@@ -506,6 +525,7 @@ impl NymVpnService {
             state_machine_shutdown_token,
             gateway_cache_handle,
             gateway_cache_join_handle,
+            discovery_refresher_join_handle,
             sentry_enabled: parameters.sentry_enabled,
             network_statistics_enabled: parameters.netstats_enabled,
             statistics_event_sender,
@@ -568,15 +588,19 @@ impl NymVpnService {
         self.services_shutdown_token.cancel();
 
         if let Err(e) = self.account_controller_handle.await {
-            tracing::error!("Failed to join on account controller handle: {}", e);
+            tracing::error!("Failed to join on account controller handle: {e}");
         }
 
         if let Err(e) = self.statistics_controller_handle.await {
-            tracing::error!("Failed to join on statistics controller handle: {}", e);
+            tracing::error!("Failed to join on statistics controller handle: {e}");
         }
 
         if let Err(e) = self.gateway_cache_join_handle.await {
-            tracing::error!("Failed to join on gateway cache handle: {}", e);
+            tracing::error!("Failed to join on gateway cache handle: {e}");
+        }
+
+        if let Err(e) = self.discovery_refresher_join_handle.await {
+            tracing::error!("Failed to join on discovery refresher handle: {e}");
         }
 
         tracing::info!("Exiting vpn service run loop");